+++ src/main/java/com/bootcamp/dto/EmployeeInformationDTO.java
@@ -9,10 +9,10 @@ public class EmployeeInformationDTO {
 	private String email;
 	private String phone_number;
 	private String hire_date;
-	private Integer salary;
+	private Double salary;
 	private String job_id;
 	private Integer manager_id;
-	private Integer commission_pct;
+	private Double commission_pct;
 	private String city;
 	private String street_address;
 	
@@ -43,7 +43,7 @@ public class EmployeeInformationDTO {
 	public String getHire_date() {
 		return hire_date;
 	}
-	public Integer getSalary() {
+	public Double getSalary() {
 		return salary;
 	}
 	public String getJob_id() {
@@ -52,7 +52,7 @@ public class EmployeeInformationDTO {
 	public Integer getManager_id() {
 		return manager_id;
 	}
-	public Integer getCommission_pct() {
+	public Double getCommission_pct() {
 		return commission_pct;
 	}
 	public String getCity() {
@@ -79,7 +79,7 @@ public class EmployeeInformationDTO {
 	public void setHire_date(String hire_date) {
 		this.hire_date = hire_date;
 	}
-	public void setSalary(Integer salary) {
+	public void setSalary(Double salary) {
 		this.salary = salary;
 	}
 	public void setJob_id(String job_id) {
@@ -88,7 +88,7 @@ public class EmployeeInformationDTO {
 	public void setManager_id(Integer manager_id) {
 		this.manager_id = manager_id;
 	}
-	public void setCommission_pct(Integer commission_pct) {
+	public void setCommission_pct(Double commission_pct) {
 		this.commission_pct = commission_pct;
 	}
 	public void setCity(String city) {
